+++ sdk/cosmos/azure-cosmos/azure/cosmos/_global_endpoint_manager.py
@@ -54,6 +54,7 @@ class _GlobalEndpointManager(object):
             client.connection_policy.UseMultipleWriteLocations,
             self.refresh_time_interval_in_ms,
         )
+        self.health_check = True
         self.refresh_needed = False
         self.refresh_lock = threading.RLock()
         self.last_refresh_time = 0
@@ -89,12 +90,17 @@ class _GlobalEndpointManager(object):
     def can_use_multiple_write_locations(self, request):
         return self.location_cache.can_use_multiple_write_locations_for_request(request)
 
-    def force_refresh(self, database_account):
+    def force_refresh_on_startup(self, database_account):
         self.refresh_needed = True
+        self.health_check = False
         self.refresh_endpoint_list(database_account)
 
     def update_location_cache(self):
         self.location_cache.update_location_cache()
+
+    def endpoint_health_check(self, database_account, **kwargs):
+        self.health_check = True
+        self.refresh_endpoint_list(database_account, **kwargs)
 
     def refresh_endpoint_list(self, database_account, **kwargs):
         if self.location_cache.current_time_millis() - self.last_refresh_time > self.refresh_time_interval_in_ms:
@@ -118,10 +124,12 @@ class _GlobalEndpointManager(object):
             if self.location_cache.should_refresh_endpoints() or self.refresh_needed:
                 self.refresh_needed = False
                 self.last_refresh_time = self.location_cache.current_time_millis()
-                database_account = self._GetDatabaseAccount(**kwargs)
-                self.location_cache.perform_on_database_account_read(database_account)
-                # this will perform getDatabaseAccount calls to check endpoint health
-                self._endpoints_health_check(**kwargs)
+                if self.health_check:
+                    # this will perform getDatabaseAccount calls to check endpoint health
+                    self._endpoints_health_check(**kwargs)
+                else:
+                    database_account, _ = self._GetDatabaseAccount(**kwargs)
+                    self.location_cache.perform_on_database_account_read(database_account)
 
     def _GetDatabaseAccount(self, **kwargs):
         """Gets the database account.
@@ -135,7 +143,7 @@ class _GlobalEndpointManager(object):
         try:
             database_account = self._GetDatabaseAccountStub(self.DefaultEndpoint, **kwargs)
             self._database_account_cache = database_account
-            return database_account
+            return database_account, self.DefaultEndpoint
         # If for any reason(non-globaldb related), we are not able to get the database
         # account from the above call to GetDatabaseAccount, we would try to get this
         # information from any of the preferred locations that the user might have
@@ -148,31 +156,38 @@ class _GlobalEndpointManager(object):
                 try:
                     database_account = self._GetDatabaseAccountStub(locational_endpoint, **kwargs)
                     self._database_account_cache = database_account
-                    return database_account
+                    return database_account, locational_endpoint
                 except (exceptions.CosmosHttpResponseError, AzureError):
                     pass
             raise
 
-    def _endpoints_health_check(self, **kwargs):
+    async def _endpoints_health_check(self, **kwargs):
         """Gets the database account for each endpoint.
 
         Validating if the endpoint is healthy else marking it as unavailable.
         """
-        all_endpoints = [self.location_cache.read_regional_endpoints[0]]
-        all_endpoints.extend(self.location_cache.write_regional_endpoints)
+        endpoints_attempted = set()
+        database_account, endpoint = await self._GetDatabaseAccount(**kwargs)
+        endpoints_attempted.add(endpoint)
+        self.location_cache.perform_on_database_account_read(database_account)
+        # should use the regions in the order returned from gateway
+        all_endpoints = [self.location_cache.available_read_regional_endpoints_by_location.values()[0]]
+        all_endpoints.extend(self.location_cache.available_write_regional_endpoints_by_location.values())
         count = 0
         for endpoint in all_endpoints:
-            count += 1
-            if count > 3:
-                break
-            try:
-                self.Client._GetDatabaseAccountCheck(endpoint.get_current(), **kwargs)
-            except (exceptions.CosmosHttpResponseError, AzureError):
-                if endpoint in self.location_cache.read_regional_endpoints:
-                    self.mark_endpoint_unavailable_for_read(endpoint.get_current(), False)
-                if endpoint in self.location_cache.write_regional_endpoints:
-                    self.mark_endpoint_unavailable_for_write(endpoint.get_current(), False)
-                    endpoint.swap()
+            if endpoint not in endpoints_attempted:
+                endpoints_attempted.add(endpoint)
+                count += 1
+                if count > 3:
+                    break
+                try:
+                    self.Client._GetDatabaseAccountCheck(endpoint.get_primary(), **kwargs)
+                    self.location_cache.mark_endpoint_available(endpoint.get_primary())
+                except (exceptions.CosmosHttpResponseError, AzureError):
+                    if endpoint in self.location_cache.read_regional_endpoints:
+                        self.mark_endpoint_unavailable_for_read(endpoint.get_primary(), False)
+                    if endpoint in self.location_cache.write_regional_endpoints:
+                        self.mark_endpoint_unavailable_for_write(endpoint.get_primary(), False)
         self.location_cache.update_location_cache()
 
     def _GetDatabaseAccountStub(self, endpoint, **kwargs):